% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ave.R
\name{meta.ave.stdmean2}
\alias{meta.ave.stdmean2}
\title{Confidence interval for an average standardized mean difference
from 2-group studies}
\usage{
meta.ave.stdmean2(alpha, m1, m2, sd1, sd2, n1, n2, stdzr, bystudy = TRUE)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{m1}{vector of estimated means for group 1}

\item{m2}{vector of estimated means for group 2}

\item{sd1}{vector of estimated SDs for group 1}

\item{sd2}{vector of estimated SDs for group 2}

\item{n1}{vector of group 1 sample sizes}

\item{n2}{vector of group 2 sample sizes}

\item{stdzr}{\itemize{
\item set to 0 for square root unweighted average variance standardizer
\item set to 1 for group 1 SD standardizer
\item set to 2 for group 2 SD standardizer
\item set to 3 for square root weighted average variance standardizer
}}

\item{bystudy}{logical to also return each study estimate (TRUE) or not}
}
\value{
Returns a matrix.  The first row is the average estimate across all studies.  If bystudy
is TRUE, there is 1 additional row for each study.  The matrix has the following columns:
\itemize{
\item Estimate - estimated effect size
\item SE - standard error
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
Computes the estimate, standard error, and confidence interval for an
average standardized mean difference from two or more 2-group studies.
Square root unweighted variances, square root weighted variances, and
single group standard deviation are options for the standardizer.
Equality of variances within or across studies is not assumed.

For more details, see Chapter 2 of Bonett (2021, Volume 5).
}
\examples{
m1 <- c(21.9, 23.1, 19.8)
m2 <- c(16.1, 17.4, 15.0)
sd1 <- c(3.82, 3.95, 3.67)
sd2 <- c(3.21, 3.30, 3.02)
n1 <- c(40, 30, 24)
n2 <- c(40, 28, 25)
meta.ave.stdmean2(.05, m1, m2, sd1, sd2, n1, n2, 0, bystudy = TRUE)

# Should return: 
#         Estimate      SE     LL     UL
# Average   1.5261 0.17343 1.1862 1.8661
# Study 1   1.6439 0.26290 1.1286 2.1592
# Study 2   1.5661 0.30563 0.9671 2.1652
# Study 3   1.4283 0.32892 0.7836 2.0729

m1 <- c(41.2, 43.2, 49.1, 40.8)
m2 <- c(36.4, 37.1, 35.9, 31.4)
sd1 <- c(4.92, 4.75, 4.87, 5.01)
sd2 <- c(4.35, 4.24, 4.12, 4.87)
n1 <- c(42, 58, 62, 39)
n2 <- c(67, 70, 84, 45)
meta.ave.stdmean2(.05, m1, m2, sd1, sd2, n1, n2, 3, bystudy = FALSE)

# Should return: 
#         Estimate      SE     LL     UL
# Average   1.8078 0.11648 1.5795 2.0361


}
\references{
\insertRef{Bonett2009a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
