% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_comp.R
\name{meta.lc.agree}
\alias{meta.lc.agree}
\title{Confidence interval for a linear contrast of G-index coefficients}
\usage{
meta.lc.agree(alpha, f11, f12, f21, f22, v)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{f11}{vector of frequency counts in cell 1,1}

\item{f12}{vector of frequency counts in cell 1,2}

\item{f21}{vector of frequency counts in cell 2,1}

\item{f22}{vector of frequency counts in cell 2,2}

\item{v}{vector of contrast coefficients}
}
\value{
Returns 1-row matrix with the following columns:
\itemize{
\item Estimate - estimated linear contrast
\item SE - standard error
\item LL - lower limit of the adjusted Wald confidence interval
\item UL - upper limit of the adjusted Wald confidence interval
}
}
\description{
Computes the estimate, standard error, and adjusted Wald confidence
interval for a linear contrast of G-index of agreement coefficients
from two or more studies. This function assumes that two raters each
provide a dichotomous rating for a sample of objects.

For more details, see Setion 3.2 of Bonett (2021, Volume 5).
}
\examples{
f11 <- c(43, 56, 49)
f12 <- c(7, 2, 9)
f21 <- c(3, 5, 5)
f22 <- c(37, 54, 39)
v <- c(.5, .5, -1)
meta.lc.agree(.05, f11, f12, f21, f22, v)

# Should return:
#          Estimate      SE      L     UL
# Contrast   0.1023 0.07972 -0.054 0.2585


}
\references{
\insertRef{Bonett2022}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
