% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_comp.R
\name{meta.lc.meanratio.ps}
\alias{meta.lc.meanratio.ps}
\title{Confidence interval for a log-linear contrast of mean ratios from
paired-samples studies}
\usage{
meta.lc.meanratio.ps(alpha, m1, m2, sd1, sd2, cor, n, v)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{m1}{vector of estimated means for measurement 1}

\item{m2}{vector of estimated means for measurement 2}

\item{sd1}{vector of estimated SDs for measurement 1}

\item{sd2}{vector of estimated SDs for measurement 2}

\item{cor}{vector of estimated correlations for paired measurements}

\item{n}{vector of sample sizes}

\item{v}{vector of contrast coefficients}
}
\value{
Returns 1-row matrix with the following columns:
\itemize{
\item Estimate - estimated log-linear contrast
\item SE - standard error of log-linear contrast
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
\item exp(Estimate) - exponentiated log-linear contrast
\item exp(LL) - lower limit of the exponentiated confidence interval
\item exp(UL) - upper limit of the exponentiated confidence interval
\item df - degrees of freedom
}
}
\description{
Computes the estimate, standard error, and confidence interval for a
log-linear contrast of paired-sample mean ratios from two or more studies.
A Satterthwaite adjustment to the degrees of freedom is used to improve
the accuracy of the confidence interval. Equality of variances within or across
studies is not assumed.

For more details, see Section 3.2 of Bonett (2021, Volume 5).
}
\examples{
m1 <- c(53, 60, 53, 57)
m2 <- c(55, 62, 58, 61)
sd1 <- c(4.1, 4.2, 4.5, 4.0)
sd2 <- c(4.2, 4.7, 4.9, 4.8)
cor <- c(.72, .78, .81, .85)
n <- c(30, 50, 30, 70)
v <- c(.5, .5, -.5, -.5)
meta.lc.meanratio.ps(.05, m1, m2, sd1, sd2, cor, n, v)

# Should return:
#           Estimate       SE         LL         UL exp(Estimate)
# Contrast 0.0440713 0.008265 0.02767047 0.06047213      1.045057
#           exp(LL)  exp(UL)    df
# Contrast 1.028057 1.062338 98.38


}
\references{
\insertRef{Bonett2020}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
