% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_comp.R
\name{meta.sub.spear}
\alias{meta.sub.spear}
\title{Confidence interval for a subgroup difference in average
Spearman correlations}
\usage{
meta.sub.spear(alpha, n, cor, group)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{n}{vector of sample sizes}

\item{cor}{vector of estimated Spearman correlations}

\item{group}{vector of group indicators:
\itemize{
\item 1 for set A
\item 2 for set B
\item 0 to ignore
}}
}
\value{
Returns a matrix with three rows:
\itemize{
\item Row 1 - estimate for Set A
\item Row 2 - estimate for Set B
\item Row 3 - estimate for difference, Set A - Set B
}

The columns are:
\itemize{
\item Estimate - estimated average correlation or difference
\item SE - standard error
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
Computes the estimate, standard error, and confidence interval
for a  difference in average Spearman correlations for two
mutually exclusive subgroups of studies. Each subgroup can have
one or more studies.

For more details, see Section 3.3 of Bonett (2021, Volume 5).
}
\examples{
n <- c(55, 190, 65, 35)
cor <- c(.40, .65, .60, .45)
group <- c(1, 1, 2, 0)
meta.sub.spear(.05, n, cor, group)

# Should return:
#                Estimate      SE      LL     UL
# Set A:            0.525 0.06484  0.3866 0.6403
# Set B:            0.600 0.08829  0.3992 0.7459
# Set A - Set B:   -0.075 0.10954 -0.2761 0.1565


}
\references{
\insertRef{Bonett2008a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
