% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rep.R
\name{replicate.gen}
\alias{replicate.gen}
\title{Compares and combines effect sizes in original and follow-up studies}
\usage{
replicate.gen(alpha, est1, se1, est2, se2)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{est1}{estimated effect size in original study}

\item{se1}{effect size standard error in original study}

\item{est2}{estimated effect size in follow-up study}

\item{se2}{effect size standard error in follow-up study}
}
\value{
A 4-row matrix. The rows are:
\itemize{
\item Row 1 summarizes the original study
\item Row 2 summarizes the follow-up study
\item Row 3 estimates the difference in effect sizes
\item Row 4 estimates the average effect size
}

Columns are:
\itemize{
\item Estimate - effect size estimate (single study, difference, average)
\item SE - standard error
\item z - z-value
\item p - p-value
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
This function can be used to compare and combine any effect size using the
effect size estimate and its standard error from the original study and
the follow-up study. The confidence level for the difference is 1 – 2*alpha,
which is recommended for equivalence testing.

For more details, see Chapter 4 of Bonett (2021, Volume 5).
}
\examples{
replicate.gen(.05, .782, .210, .650, .154)

# Should return: 
#                       Estimate        SE     z     p         LL        UL
# Original:                0.782 0.2100000 3.724 0.000  0.3704076 1.1935924
# Follow-up:               0.650 0.1540000 4.221 0.000  0.3481655 0.9518345
# Original - Follow-up:    0.132 0.2604151 0.507 0.612 -0.2963446 0.5603446
# Average:                 0.716 0.1302075 5.499 0.000  0.4607979 0.9712021


}
\references{
\insertRef{Bonett2021a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
