% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rep.R
\name{replicate.meanratio.ps}
\alias{replicate.meanratio.ps}
\title{Compares and combines paired-samples mean ratios in original and follow-up
studies}
\usage{
replicate.meanratio.ps(
  alpha,
  m11,
  m12,
  sd11,
  sd12,
  cor1,
  n1,
  m21,
  m22,
  sd21,
  sd22,
  cor2,
  n2
)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{m11}{estimated mean for group 1 in original study}

\item{m12}{estimated mean for group 2 in original study}

\item{sd11}{estimated SD for group 1 in original study}

\item{sd12}{estimated SD for group 2 in original study}

\item{cor1}{estimated correlation of paired measurements in original study}

\item{n1}{sample size in original study}

\item{m21}{estimated mean for group 1 in follow-up study}

\item{m22}{estimated mean for group 2 in follow-up study}

\item{sd21}{estimated SD for group 1 in follow-up study}

\item{sd22}{estimated SD for group 2 in follow-up study}

\item{cor2}{estimated correlation of paired measurements in follow-up study}

\item{n2}{sample size in follow-up study}
}
\value{
A 4-row matrix. The rows are:
\itemize{
\item Row 1 summarizes the original study
\item Row 2 summarizes the follow-up study
\item Row 3 estimates the difference in mean differences
\item Row 4 estimates the average mean difference
}

The columns are:
\itemize{
\item Estimate - mean difference estimate (single study, difference, average)
\item SE - standard error
\item t - t-value
\item p - p-value
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
\item df - degrees of freedom
}
}
\description{
This function computes confidence intervals from an original study and a
follow-up study where the effect size is a paired-samples mean ratio.
Confidence intervals for the log-difference and average log effect size are
also computed. Equality of variances within or across studies is not assumed.
A Satterthwaite adjustment to the degrees of freedom is used to improve the
accuracy of the confidence intervals. The confidence level for the difference
is 1 – 2*alpha, which is recommended for equivalence testing.

For more details, see Chapter 4 of Bonett (2021, Volume 5).
}
\examples{
replicate.meanratio.ps(.05, 86.22, 70.93, 14.89, 12.32, .765, 20, 
                       84.81, 77.24, 15.68, 16.95, .702, 75)

# Should return:
#                         Estimate         SE       t       p    df
# Original:             0.30766736 0.02420564 12.7106 0.00000 39.00
# Follow-up:            0.27715566 0.01532870 18.0808 0.00000 74.00
# Original - Follow-up: 0.03051171 0.02865104  1.0649 0.29055 70.57
# Average:              0.29241151 0.01432552 20.4119 0.00000 70.57
#                       exp(Estimate)   exp(LL)  exp(UL)
# Original:                  1.360248 1.2952540 1.428504
# Follow-up:                 1.319372 1.2796832 1.360291
# Original - Follow-up:      1.030982 0.9829058 1.081410
# Average:                   1.339654 1.3019254 1.378476


}
\references{
\insertRef{Bonett2020}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
