% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rep.R
\name{replicate.oddsratio}
\alias{replicate.oddsratio}
\title{Compares and combines odds ratios in original and follow-up studies}
\usage{
replicate.oddsratio(alpha, est1, se1, est2, se2)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{est1}{estimate of log odds ratio in original study}

\item{se1}{standard error of log odds ratio in original study}

\item{est2}{estimate of log odds ratio in follow-up study}

\item{se2}{standard error of log odds ratio in follow-up study}
}
\value{
A 4-row matrix. The rows are:
\itemize{
\item Row 1 summarizes the original study
\item Row 2 summarizes the follow-up study
\item Row 3 estimates the ratio of odds ratios
\item Row 4 estimates the geometric average odds ratio
}

The columns are:
\itemize{
\item Estimate - log odds ratio estimate (single study, ratio, average)
\item SE - standard error of log odds estimate
\item z - z-value
\item p - p-value
\item exp(Estimate) - exponentiated estimate
\item exp(LL) - exponentiated lower limit of the confidence interval
\item exp(UL) - exponentiated upper limit of the confidence interval
}
}
\description{
This function computes confidence intervals for an odds ratio from an
original study and a follow-up study. Confidence intervals for the
ratio of odds ratios and geometric average odds ratio are also
computed. The confidence level for the ratio of ratios is 1 – 2*alpha, which
is recommended for equivalence testing.

For more details, see Chapter 4 of Bonett (2021, Volume 5).
}
\examples{
replicate.oddsratio(.05, 1.39, .302, 1.48, .206)

# Should return:
#                        Estimate        SE      z     p
# Original:            1.39000000 0.3020000  4.603 0.000
# Follow-up:           1.48000000 0.2060000  7.184 0.000
# Original/Follow-up: -0.06273834 0.3655681 -0.172 0.864
# Average:             0.36067292 0.1827840  1.973 0.048
#                     exp(Estimate)   exp(LL)  exp(UL)
# Original:               4.0148501 2.2212961 7.256583
# Follow-up:              4.3929457 2.9336501 6.578144
# Original/Follow-up:     0.9391892 0.5147653 1.713551
# Average:                1.4342943 1.0024257 2.052222


}
\references{
\insertRef{Bonett2021a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
