% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rep.R
\name{replicate.propratio2}
\alias{replicate.propratio2}
\title{Compares and combines 2-group proportion ratios in original and follow-up
studies}
\usage{
replicate.propratio2(alpha, f11, f12, n11, n12, f21, f22, n21, n22)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{f11}{frequency count for group 1 in original study}

\item{f12}{frequency count for group 2 in original study}

\item{n11}{sample size for group 1 in original study}

\item{n12}{sample size for group 2 in original study}

\item{f21}{frequency count for group 1 in follow-up study}

\item{f22}{frequency count for group 2 in follow-up study}

\item{n21}{sample size for group 1 in follow-up study}

\item{n22}{sample size for group 2 in follow-up study}
}
\value{
A 4-row matrix. The rows are:
\itemize{
\item Row 1 summarizes the original study
\item Row 2 summarizes the follow-up study
\item Row 3 estimates the ratio of proportion ratios
\item Row 4 estimates the geometric average proportion ratio
}

The columns are:
\itemize{
\item Estimate - log proportion ratio estimate (single study, ratio, average)
\item SE - standard error of log proportion estimate
\item z - z-value
\item p - p-value
\item exp(Estimate) - exponentiated estimate
\item exp(LL) - exponentiated lower limit of the confidence interval
\item exp(UL) - exponentiated upper limit of the confidence interval
}
}
\description{
This function computes confidence intervals from an original study and a
follow-up study where the effect size is a 2-group log proportion ratio.
Confidence intervals for exponentiated effect sizes are also computed.
The confidence level for the ratio of ratios is 1 – 2*alpha,
which is recommended for equivalence testing.

For more details, see Chapter 4 of Bonett (2021, Volume 5).
}
\examples{
replicate.propratio2(.05, 21, 16, 40, 40, 19, 13, 60, 60)

# Should return:
#                       Estimate        SE      z     p
# Original:            0.2682640 0.2463597  1.089 0.276
# Follow-up:           0.3735135 0.3082711  1.212 0.226
# Original/Follow-up: -0.1052495 0.3946190 -0.267 0.789
# Average:             0.3208887 0.1973095  1.626 0.104
#                     exp(Estimate)   exp(LL)  exp(UL)
# Original:               1.3076923 0.8068705 2.119373
# Follow-up:              1.4528302 0.7939881 2.658372
# Original/Follow-up:     0.9000999 0.4703209 1.722611
# Average:                1.3783522 0.9362893 2.029132


}
\references{
\insertRef{Bonett2021}{vcmeta}
}
