% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_se.R
\name{se.median}
\alias{se.median}
\title{Computes the standard error for a median}
\usage{
se.median(alpha, LL, UL, n, type)
}
\arguments{
\item{alpha}{alpha value for 1-alpha confidence interval}

\item{LL}{lower limit of confidence interval}

\item{UL}{upper limit of confidence interval}

\item{n}{sample size}

\item{type}{\itemize{
\item set to 1 for classical confidence interval
\item set to 2 for any other type of confidence interval
}}
}
\value{
Returns the estimated standard error of the sample median
}
\description{
Computes the standard error of a median using a confidence interval for the
distribution-free interval (see Snedecor & Cochran, 1980), this function
computes a Price-Bonett standard error which is known to be very accurate.
If any other type of confidence interval is used, this function computes an
approximate standard error.

In a 2-group design, this function can be used to compute the standard error
of the median in each group. Then the difference in estimated medians in the
two groups can be used as the effect size and the standard error of the
difference can be set to the square root of the sum of the squared standard
errors from each group. Single-group medians or median differences and their
standard errors can be used as input in the \link[vcmeta]{meta.ave.gen},
\link[vcmeta]{meta.lc.gen}, and and \link[vcmeta]{meta.lm.gen} functions.
}
\examples{
se.median(.05, 47.21, 68.68, 35, 1)

# Should return:
#        SE
#  5.252114


}
\references{
\insertRef{Price2001}{vcmeta}

\insertRef{Snedecor1980}{vcmeta}
}
