% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_se.R
\name{se.stdmean.ps}
\alias{se.stdmean.ps}
\title{Computes the standard error for a paired-samples standardized mean
difference}
\usage{
se.stdmean.ps(m1, m2, sd1, sd2, cor, n, stdzr)
}
\arguments{
\item{m1}{estimated mean for measurement 1}

\item{m2}{estimated mean for measurement 2}

\item{sd1}{estimated standard deviation for measurement 1}

\item{sd2}{estimated standard deviation for measurement 2}

\item{cor}{estimated correlation for measurements 1 and 2}

\item{n}{sample size}

\item{stdzr}{\itemize{
\item set to 0 for square root average variance standardizer
\item set to 1 for measurement 1 SD standardizer
\item set to 2 for measurement 2 SD standardizer
}}
}
\value{
Returns a one-row matrix:
\itemize{
\item Estimate - estimated standardized mean difference
\item SE - standard error
}
}
\description{
Computes the standard error of a paired-samples standardized mean
difference using the sample size, estimated means, estimated standard
deviations, and estimated correlation. The effect size estimate and standard error
output from this function can be used as input in the \link[vcmeta]{meta.ave.gen},
\link[vcmeta]{meta.lc.gen}, and \link[vcmeta]{meta.lm.gen} functions in
applications where compatible standardized mean differences from a combination
of 2-group and paired-samples experiments are used in the meta-analysis.
Equality of variances is not assumed.

For more details, see Chapter 1 of Bonett (2021, Volume 5)
}
\examples{
se.stdmean.ps(23.9, 25.1, 1.76, 2.01, .78, 25, 0)

# Should return: 
#                                 Estimate      SE
# Standardized mean difference:   -0.6352 0.16029

se.stdmean.ps(23.9, 25.1, 1.76, 2.01, .78, 25, 1)

# Should return:
#                                Estimate      SE
# Standardized mean difference:   -0.6818 0.17738


}
\references{
\insertRef{Bonett2009a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
