% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ave.R
\name{meta.ave.cronbach}
\alias{meta.ave.cronbach}
\title{Confidence interval for an average Cronbach alpha reliability}
\usage{
meta.ave.cronbach(alpha, n, rel, r, bystudy = TRUE)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{n}{vector of sample sizes}

\item{rel}{vector of sample reliabilities}

\item{r}{number of measurements (e.g., items) used to compute each reliability}

\item{bystudy}{logical to also return each study estimate (TRUE) or not}
}
\value{
Returns a matrix.  The first row is the average estimate across all studies.  If bystudy
is TRUE, there is 1 additional row for each study.  The matrix has the following columns:
\itemize{
\item Estimate - estimated effect size
\item SE - standard error
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
Computes the estimate, standard error, and confidence interval for an
average Cronbach reliability coefficient from two or more studies.

For more details, see Chapter 2 of Bonett (2021, Volume 5).
}
\examples{
n <- c(583, 470, 546, 680)
rel <- c(.91, .89, .90, .89)
meta.ave.cronbach(.05, n, rel, 10, bystudy = TRUE)

# Should return:
#         Estimate      SE     LL     UL
# Average   0.8975 0.00326 0.8911 0.9039
# Study 1   0.9100 0.00557 0.8986 0.9204
# Study 2   0.8900 0.00758 0.8744 0.9041
# Study 3   0.9000 0.00639 0.8869 0.9119
# Study 4   0.8900 0.00630 0.8771 0.9018


}
\references{
\insertRef{Bonett2010}{vcmeta}

\insertRef{Bonett2015b}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
