% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rep.R
\name{replicate.spear}
\alias{replicate.spear}
\title{Compares and combines Spearman correlations in original and follow-up studies}
\usage{
replicate.spear(alpha, cor1, n1, cor2, n2)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{cor1}{estimated Spearman correlation in original study}

\item{n1}{sample size in original study}

\item{cor2}{estimated Spearman correlation in follow-up study}

\item{n2}{sample size in follow-up study}
}
\value{
A 4-row matrix. The rows are:
\itemize{
\item Row 1 summarizes the original study
\item Row 2 summarizes the follow-up study
\item Row 3 estimates the difference in correlations
\item Row 4 estimates the average correlation
}

The columns are:
\itemize{
\item Estimate - Spearman correlation estimate (single study, difference, average)
\item SE - standard error
\item z - z-value
\item p - p-value
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
This function can be used to compare and combine Spearman correlations from
an original study and a follow-up study. The confidence level for the
difference is 1 – 2*alpha, which is recommended for equivalence testing.

For more details, see Chapter 4 of Bonett (2021, Volume 5).
}
\examples{
replicate.spear(.05, .598, 80, .324, 200)

# Should return:
#                       Estimate      SE     z     p     LL     UL
# Original:                0.598 0.07948 5.315 0.000 0.4199 0.7318
# Follow-up:               0.324 0.06542 4.571 0.000 0.1905 0.4457
# Original - Follow-up:    0.274 0.10294 3.438 0.001 0.0948 0.4342
# Average:                 0.461 0.05147 9.968 0.000 0.3670 0.5457


}
\references{
\insertRef{Bonett2021a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
