% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rep.R
\name{replicate.stdmean2}
\alias{replicate.stdmean2}
\title{Compares and combines 2-group standardized mean differences in original and
follow-up studies}
\usage{
replicate.stdmean2(
  alpha,
  m11,
  m12,
  sd11,
  sd12,
  n11,
  n12,
  m21,
  m22,
  sd21,
  sd22,
  n21,
  n22,
  stdzr
)
}
\arguments{
\item{alpha}{alpha level for 1-alpha confidence}

\item{m11}{estimated mean for group 1 in original study}

\item{m12}{estimated mean for group 2 in original study}

\item{sd11}{estimated SD for group 1 in original study}

\item{sd12}{estimated SD for group 2 in original study}

\item{n11}{sample size for group 1 in original study}

\item{n12}{sample size for group 2 in original study}

\item{m21}{estimated mean for group 1 in follow-up study}

\item{m22}{estimated mean for group 2 in follow-up study}

\item{sd21}{estimated SD for group 1 in follow-up study}

\item{sd22}{estimated SD for group 2 in follow-up study}

\item{n21}{sample size for group 1 in follow-up study}

\item{n22}{sample size for group 2 in follow-up study}

\item{stdzr}{\itemize{
\item set to 0 for square root unweighted average variance standardizer
\item set to 1 for group 1 SD standardizer
\item set to 2 for group 2 SD standardizer
\item set to 3 for square root weighted average variance standardizer
}}
}
\value{
A 4-row matrix. The rows are:
\itemize{
\item Row 1 summarizes the original study
\item Row 2 summarizes the follow-up study
\item Row 3 estimates the difference in standardized mean differences
\item Row 4 estimates the average standardized mean difference
}

The columns are:
\itemize{
\item Estimate - standardized mean difference estimate (single study, difference, average)
\item SE - standard error
\item LL - lower limit of the confidence interval
\item UL - upper limit of the confidence interval
}
}
\description{
This function computes confidence intervals from an original study and a
follow-up study where the effect size is a 2-group standardized mean
difference. Confidence intervals for the difference and average effect
size are also computed. Equality of variances within or across studies
is not assumed. The confidence level for the difference is 1 – 2*alpha,
which is recommended for equivalence testing. Square root unweighted
variances, square root weighted variances, and single-group standard
deviation are options for the standardizer.

For more details, see Chapter 4 of Bonett (2021, Volume 5).
}
\examples{
replicate.stdmean2(.05, 21.9, 16.1, 3.82, 3.21, 40, 40, 
                        25.2, 19.1, 3.98, 3.79, 75, 75, 0)

# Should return: 
#                        Estimate     SE      LL     UL
#  Original:               1.6280 0.2595  1.1353 2.1524
#  Follow-up:              1.5617 0.1871  1.2030 1.9363
#  Original - Follow-up:   0.0742 0.3199 -0.4519 0.6004
#  Average:                1.5949 0.1599  1.2814 1.9083


}
\references{
\insertRef{Bonett2021a}{vcmeta}

\insertRef{Bonett2021}{vcmeta}
}
