% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_se.R
\name{se.cohen}
\alias{se.cohen}
\title{Computes the standard error for Cohen's d}
\usage{
se.cohen(d, n1, n2)
}
\arguments{
\item{d}{estimated Cohen's d}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}
}
\value{
Returns a one-row matrix:
\itemize{
\item Estimate - Cohen's d (from input)
\item SE - standard error
}
}
\description{
Computes the standard error of Cohen's d using only the two
sample sizes and an estimate of Cohen's d. Cohen's d and its standard error
assume equal variances. The estimate of Cohen's d, with the standard error
output from this function, can be used as input in the \link[vcmeta]{meta.ave.gen},
\link[vcmeta]{meta.lc.gen}, and \link[vcmeta]{meta.lm.gen} functions in
applications where different types of compatible standardized mean
differences are used in the meta-analysis. If the means, standard deviations,
and sample sizes for the two groups are available, use the
\link[vcmeta]{se.stdmean2} function which does not assume equal variances. The
standard error for Cohen's d can be very inaccurate if the variances are
unequal and the sample sizes are unequal.

For more details, see Chapter 1 of Bonett (2021, Volume 5)
}
\examples{
se.cohen(.782, 35, 50)

# Should return: 
#            Estimate      SE
# Cohen's d:    0.782 0.22887


}
\references{
\insertRef{Bonett2021}{vcmeta}
}
