% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_se.R
\name{se.spear}
\alias{se.spear}
\title{Computes the standard error for a Spearman correlation}
\usage{
se.spear(cor, n)
}
\arguments{
\item{cor}{estimated Spearman correlation}

\item{n}{sample size}
}
\value{
Returns a one-row matrix:
\itemize{
\item Estimate - estimate of Spearman correlation
\item SE - standard error
}
}
\description{
This function can be used to compute the standard error of a
Spearman correlation using the estimated correlation and sample
size. The standard error from this function can be used as input
in the \link[vcmeta]{meta.ave.gen}, \link[vcmeta]{meta.lc.gen}, and
\link[vcmeta]{meta.lm.gen} functions in applications where a combination
of different types of correlations are used in the meta-analysis.
}
\examples{
se.spear(.40, 55)

# Should return: 
#                       Estimate        SE
# Spearman correlation:      0.4 0.1210569


}
\references{
\insertRef{Bonett2000}{vcmeta}
}
